factor out import cycle in appsec
